+++ README.md
@@ -24,3 +24,13 @@ Napisz REST API dla listy pojazdów.
 
 ## Dodatkowo
 - [x] walidacja danych wejściowych
+
+#M4_HomeWork_Thymeleaf
+Rozbuduj swoją aplikacje z poprzedniego tygodnia o interfejs graficzny, który umożliwi:
+
+- [ ] wyświetlanie wszystkich pozycji
+- [ ] pobieranie elementu po jego id
+- [ ] dodawanie pozycji
+- [ ] modyfikowanie pozycji
+- [ ] modyfikowanie jednego z pól pozycji
+- [ ] usuwania jednej pozycji
