+++ src/main/resources/templates/add-car.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>AddCar</title>
+</head>
+<body>
+<form th:action="@{/add-car}" th:object="${newCar}" method="post">
+    <p>Id : <input type="text" th:field="*{id}"></p>
+    <p>Mark: <input type="text" th:field="*{mark}"></p>
+    <p>Model: <input type="text" th:field="*{model}"></p>
+    <p>Color: <input type="text" th:field="*{color}"></p>
+    <p><input type="submit" value="Add new car"></p>
+</form>
+
+</body>
+</html>
